Splint fiddles.
CVS patchset: 7435
CVS date: 2004/10/09 17:29:22

+++ rpmio/sexp/sexp.h
@@ -90,7 +90,7 @@ void * sexpAlloc(size_t n)
 sexpSimpleString newSimpleString(void)
 	/*@globals fileSystem @*/
 	/*@modifies fileSystem @*/;
-long int simpleStringLength(sexpSimpleString ss)
+size_t simpleStringLength(sexpSimpleString ss)
 	/*@*/;
 /*@exposed@*/ /*@null@*/
 octet *simpleStringString(sexpSimpleString ss)
